+++ flagging_site/blueprints/README.md
@@ -3,3 +3,4 @@
 - `__init__.py`: required to treat directory as package
 - `cyanobacteria.py`: 
 - `flagging.py`: render index.html and output table model using data retrieved from usgs and hobolink 
+- `api.py`: provide info and updates on api status
