+++ flagging_site/blueprints/__init__.py
@@ -1,2 +1,3 @@
 from . import cyanobacteria
 from . import flagging
+from . import api
